requirements.txt
commonmark==0.9.1
Django==4.0.5
djangorestframework==3.13.1
PyQRCode==1.2.1
